requirements.txt
annotated-types==0.7.0
anyio==4.11.0
bcrypt==5.0.0
beanie==2.0.1
certifi==2025.6.15
cffi==2.0.0
charset-normalizer==3.4.2
click==8.3.0
colorama==0.4.6
cryptography==46.0.3
distlib==0.4.0
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.118.0
filelock==3.18.0
fsspec==2025.5.1
greenlet==3.2.4
h11==0.16.0
idna==3.10
Jinja2==3.1.6
lazy-model==0.4.0
llvmlite==0.44.0
MarkupSafe==3.0.2
more-itertools==10.7.0
motor==3.7.1
mpmath==1.3.0
networkx==3.5
numba==0.61.2
numpy==2.2.6
openai-whisper==20240930
opencv-python==4.12.0.88
packaging==25.0
pandas==2.3.3
passlib==1.7.4
pillow==12.0.0
platformdirs==4.4.0
pyasn1==0.6.1
pycparser==2.23
pydantic==2.11.9
pydantic-settings==2.12.0
pydantic_core==2.33.2
PyJWT==2.10.1
pymongo==4.15.5
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.20
pytz==2025.2
regex==2024.11.6
requests==2.32.4
rsa==4.9.1
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
sqlmodel==0.0.27
starlette==0.48.0
sympy==1.14.0
tiktoken==0.9.0
torch==2.7.1
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.37.0
virtualenv==20.34.0
virtualenvwrapper-win==1.2.7
